[FEATURE] Move linkvalidator reports to new module
The linkvalidator backend module was previously
built into the Web->Info module.

Now, Linkvalidator receives its own module in the
"Web" area. This allows to define access permissions
for editors via the module access logic, and shapes
the functionality in a cleaner way.

The Linkvalidator module not only reports but also
allows to validate. Through this change, the module
is now accessible via a dedicated speaking backend URL.

This change makes the feature much more prominent
and allows for further refinements such as pagination.

Resolves: #96465
Releases: main
Change-Id: I83a4ed3f31cfbe59d61778b0d527f5f4704a404a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72701
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>