[BUGFIX] Resolve shortcuts in HMENU to access restricted pages
There is a difference between

> typolink.linkAccessRestrictedPages = 1 (only a toggle)
and
> HMENU.showAccessRestrictedPages = [pageId]|NONE

HMENU.showAccessRestrictedPages behaves like
the global option "config.typolinkLinkAccessRestrictedPages"

Basically, if a page is access restricted, link
to a different {pageId}.

This change explains the behavior:
-- https://review.typo3.org/c/Packages/TYPO3.CMS/+/35908/

"typolink.linkAccessRestrictedPages" in contrast
still links to the actual disallowed page, which
is also nice in case you have a 403 error page in place.

This change fixes the behaviour of HMENU to behave
EXACTLY like the global config.typolinkLinkAccessRestrictedPages,
previously HMENU did some magic PLUS it set
"typolink.linkAccessRestrictedPages"

With this change, shortcuts to access restricted pages
now get properly transformed and linked for menus.

Resolves: #60258
Resolves: #65118
Related: #63804
Releases: main, 11.5
Change-Id: Ifd975243fe4b024b3fcbd4e356430d809cc0f429
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72785
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>