[BUGFIX] Prevent redirects causing infinite redirect loops
This patch adds a check to detect self referencing
redirects, thus avoiding them and instead log an error
in the corresponding frontend redirect middleware.

Furthermore, add a bunch of tests along the way to cover
this change and the different constellations, for example
not avoiding redirect with the same path but external host.

Resolves: #96427
Releases: main, 11.5, 10.4
Change-Id: I554ba51b53065dd754068e379f69c2a5dffc3054
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72856
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>