[BUGFIX] Avoid duplicate pages in pagetree when rootPid is mounted
With the changes in #95854, implicit rootPid dbmounts caused
duplicated pages in pagetree, because explicit page mounts have
been considered multiple times for the pid=0 tree:

 * First time because of their relationship to pid = 0
 * Second time because of their explicit mount
   (which was sorted into the pid = 0 tree)

This happened because #95854 optimized to only use one call
to the PageTree repository, which then had the side effect
of merging "multiple" pages into one tree, instead
of showing "two" separate trees.

The patch ensures that multiple page trees are used when
the rootPid (0) is included as a webmount.

Releases: main, 11.5, 10.4
Resolves: #96367
Related: #95854
Change-Id: I7f421995f2d4aa88bb266fec684565102a4dc080
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72812
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>