[BUGFIX] Add TCA migration for select authMode keywords
With the introduction of itemGroups, the index 3 of
the select items array has been shifted one position
to the right. Before that, the index 3 was used for
descriptions and index 4 for an optional keyword
EXPL_ALLOW or EXPL_DENY. These are used together
with authMode=individual to explicitly allow or deny
single items.

Since descriptions now occupy the index 4, the
former usage of this index needs to be shifted as
well to index 5.

For backwards compatibility reasons, a TCA migration
is added, which will check for these special
keywords and move them one index up.

Resolves: #96444
Related: #91008
Releases: main, 11.5, 10.4
Change-Id: I32a96f5c6377871551ab4ee60a402a585da7eaa0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72820
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>