[BUGFIX] Use realpath() for vcsIgnore in CGL fixer
It seems that symfony >= 5.4.0 has a bug in symfony
finder that .gitignore files are no longer always
found when the ->in() directive is a non-absolute
path with '/../'.
This triggers issues with our cgl fixer script in
nightly mode.
We work around that by using an absolute path
instead in the configuration and will report the
issue to symfony.

Resolves: #96450
Related: #96171
Releases: master, 11.5
Change-Id: Ice865b92c118e26d519fbca7ae49e41fbe4d5110
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72842
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>