[BUGFIX] Mark shortcut pages as active in HMENU
This change moves the method resolveShortcutPage()
into PageRepository as it is now also used in determining
if a page is marked as active / current.

The method `resolveShortcutPage()` is much more convenient
when working with $page records instead of the properties,
instead of `getPageShortcut()`.

This also allows MenuProcessor to have a proper "active"
and "current" values set properly without any workarounds.

Resolves: #85138
Resolves: #80841
Resolves: #87923
Releases: main, 11.5
Change-Id: Ic401fb42696757199c5974120f24250c467d9b75
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72860
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>