[BUGFIX] Change misleading label for workspace mass action
With the introduction of the multi record selection, the workspace
actions, shown above and below the records, got prepended with
labels.

While the mass action targets the entire workspace, the label just
read "all", which is misleading because one could think the action
only targets all records of the current view and then accidentally
publish or even worse, discard the entire workspace.

This is fixed by changing the label.

Resolves: #96468
Releases: main, 11.5
Change-Id: Id2205e843656af7576f7038dddc81ae53308a185
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72886
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>