[TASK] Remove special handling of translated shortcuts
When a page is hit in TYPO3 v12.0, the page translation
is automatically applied nowadays. Before, this
was only possible if TypoScript was applied (finding out
the config.sys_language_uid TypoScript setting etc).

For this reason, the functionality can be removed.

Same applies for determineOriginalShortcutPage()
which is not needed anymore since "pages_language_overlay"
is migrated into "pages" and "allowLanguageSynchronization"
is active.

Resolves: #96540
Releases: main
Change-Id: I5158dc62dc5a404e14b0685f4d177fb47723cd82
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72864
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>