[BUGFIX] TypoLink is executed fully without href attribute
In #87992 an early return was added to have <a> tags
with just an "id" or "name" attribute. This is all nice,
but when used further typolink settings (such as userFunc
or custom "title"), these options were never run.

This change now uses typolink() natively again and
removes the introduced "resolveAnchorLinK" workaround.

Added tests show that additional attributes are executed
and links without href are actually created (was previously
untested).

Resolves: #96464
Related: #87992
Releases: main, 11.5
Change-Id: I67d4090228684fc89f72bc3d6367109437040ccf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72916
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>