[BUGFIX] HMENU: Use ATagParams directly from typolink
When setting "HMENU.1.NO.ATagParams" then "ATagParams"
called in cObj->ATagParams is now processed once by
setting the option in typolink().

This way, any hooks in ATagParams are added within
the context of the typolink function, also receiving
all information in the LinkResult object. This is
now possible with TYPO3 v11.

Some workarounds are needed and document on how
to clean this functionality up in TYPO3 v12+.

Resolves: #96486
Releases: main
Change-Id: I5804931ec623035247baf34a226fe70e44736a72
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72909
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>