[BUGFIX] Bring back functionality for $uniqueIds in SelectSingleElement
This resolves a todo for a missing usage of a filled
array called "$uniqueIds".

This array had two responsibilities:

1. For inline fields with "foreign_unique" set, all
already used unique ids must be excluded from the
select items.

2. Ensure a select field inside an inline field,
which functions as one (symmetric) side of a
mn-table, will not have the same field we are
currently in, in its select items. In short: This
prevents self-references.

The previous functionality has been restored with
added type-safety.

This functionality was probably removed by accident
during a refactoring process.

Resolves: #96487
Related: #64762
Releases: main, 11.5
Change-Id: I75d68fdaba5bdf896c336fe428652d15d760ac4c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74068
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>