[FEATURE] Introduce entry points for TCA type "group"
To improve the workflow for editors while selecting records
or folders in TCA type "group" fields, a new field configuration
"entryPoints" is introduced. It can be used to define
the default page / folder, which should be selected when
opening the element browser. This can be configured for
all or for each table individually.

This configuration might be quite useful since TYPO3
systems usually use dedicated storage pages for record
types nowadays.

To further support site administrators, the new configuration
also allows the usage of the known TCA markers and additionally
is also added to FormEngine's "allowOverrideMatrix".

Resolves: #91077
Releases: main
Change-Id: Ie6e4e8675ff07288480bce69467805409defdc57
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72911
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>