[BUGFIX] Calculate action menu generation earlier in PageLayoutController
With #95387 the menu generation of the page module
was moved to the end of the request handling method.

However, the `makeActionMenu()` method does currently
not only create the menu but does also make some sanity
check, whether the currently selected `function` is actually
available.

If not, the `function` is reset to the default ("columns mode").
This e.g. happens when in languages mode and then switching
to another domain / another site, which is single language only.
This then led to the user being unable to work in "columns mode",
since no action menu is generated for such sites.

This is now fixed by executing the mentioned method before the
content gets generated.

Resolves: #96084
Related: #95387
Releases: main, 11.5
Change-Id: Iadb67ece1724b6d34a4724d7782bdd295cece642
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72971
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>