[BUGFIX] Use optgroup in SelectMultipleSideBySideElement
Instead of using multiple "fake" <option> elements with
the same value `--div--`, the SelectMultipleSideBySideElement
does now use proper HTML <optgroup> elements for grouping
of the available options. This is in line with other
select elements, e.g. the SelectSingleElement.

Besides the improved HTML markup, this also improves
the UI, since the <optgroup> element is non-selectable.
It is also no longer considered by the filter, which
previously led to confusion, especially when filtering
and having the "dividers" as only options left.

Resolves: #95137
Releases: main, 11.5
Change-Id: Ia51e2623217eb0c7abd6c0cd2e9c4a742686641d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72992
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>