[TASK] Simplify EXT:redirects ManagementController
While the ManagementController does only feature
one action - overviewAction - the handleRequest()
made a couple of unnecessary calls, especially to
the mentioned action.

This is now simplified together with a couple of
further cleanups, e.g. removal of unnecessary
class properties / methods.

Additionally, the new BackendTemplateView is
now used instead of StandaloneView, which
required a few changes in the template.

Resolves: #96534
Releases: main
Change-Id: I5a5d81ad98b87083504444eb7c2bf8420227b159
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73016
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>