[TASK] Migrate Form Framework JavaScript and documentation to ES6
The module syntax is moved from AMD to ES6.
To avoid reindenting every javasscript line (>10000 LOC),
the existings factories are kept as-is and the exports
are spreaded from the factory result.

A new configuration option dynamicJavaScriptModules
is added to replace the previous dynamicRequireJsModules
functionality to incorparate for the new ES6 functionality
added in #96510 and #96511.

Also dynamicRequireJsModules is no longer passed to
the form editor and manager templates, as this variable
is a left over from #90867.

Resolves: #96568
Related: #96323
Releases: main
Change-Id: I68bf3441f34b9f365385b5ae54f8c1f4b0bca599
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73026
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>