[TASK] runTests.sh: Skip reports.xml in acceptance tests
The created reports.xml when running acceptance tests
is used for nothing in CI and just uploaded as artifact
along with the other reports. When ac tests fail, the
generally useful file is reports.html together with the
screenshots. Creating the xml file is skipped now.

Users who still want to create the xml file when running
the acceptanc test suite can do so by adding the "extra"
argument to runTests.sh:
> Build/Scripts/runTests.sh -s acceptance -e '--xml reports.xml'

Resolves: #96557
Releases: main, 11.5, 10.4
Change-Id: Ic385e307bb48702bc9269e42960fc76f1aa0be41
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73039
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>