[BUGFIX] Create module template early in PreviewController
When creating the module template, also default CSS
and JS is added (using PageRenderer).

In the PreviewController, the loading order is important,
since module related properties are overwritten in a
custom CSS file (e.g. margins).

Therefore, the module creation is now moved to the top
of the method, which allows to override the default
styles again.

Resolves: #96581
Related: #96548
Releases: main
Change-Id: I7c14d190b760e9e847f9067c34b31f0d0e0b603a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73060
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>