[BUGFIX] Use LanguageServiceFactory in FE RequestHandler
When previewing pages a null pointer exception
is triggered in the frontend RequestHandler, in
case $GLOBALS['LANG'] does not exist. This
is the case when no user object exists (or is
not a FrontendBackendUserAuthentication).

This is now fixed by creating the LanguageService
with the corresponding factory and no longer relying
on $GLOBALS['LANG'].

Additionally, another possible null pointer is
fixed in the factory, since createFromUserPreferences()
allows NULL for the first parameter "$user"
but previously did not check the provided
value before accessing.

Resolves: #96590
Releases: main, 11.5, 10.4
Change-Id: I160a9dae9131973d681b392ee713d1fadcb9a24b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73095
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>