[BUGFIX] Avoid undefined array key 'casesensitiveComp'
Using null coalescing operator with fallback to 'false'
to avoid undefined array key access warnings when key
'casesensitiveComp' is not set. Using 'false' here as
this matches the behaviour of <PHP8.0, and as this is
used as method argument of type bool a bool typecase
is added along the way to be on the safe side.

Resolves: #96537
Releases: main, 11.5
Change-Id: I3c4df09b2adf5eec57e31947189f4a1e2aea4bd2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73088
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>