[BUGFIX] Remove undefined variable in extensionmanager ListUtility
Fixing an issue with package informations to list extensions for
the cli command 'extension:list' code has been moved  #95693, which
left over an undefined variable at the new place. This has been
undetected and not failed because of a used 'null coalescing operator'
with a default value as fallback.

However well configured static code analysers may struggle with this
and emiting errors, we will remove this undefined variable here as
which should be fine. Additional enriching emconf information has
been adjusted to always include the state, which required to adjust
corresponding unit tests for it.

Resolves: #96602
Related: #95693
Releases: main, 11.5
Change-Id: I588d8ee724f7690af5759da0094719a9c287e4c2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73155
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>