[TASK] Use native TYPO3Fluid view in ext:seo frontend
To demonstrate it's possible to use the native
TYPO3 "standalone" Fluid TemplateView in the frontend,
the patch changes the ext:seo frontend sitemap XML
rendering to use TYPO3Fluid TemplateView instead of
ext:fluid StandaloneView.
This means the view does not receive a request anymore
and needs to be set up in a slightly different way,
which in turn avoids depedencies to extbase for this
non-extbase based plugin.
The extbase specific f:uri.page ViewHelper usage is
thus changed to f:uri.typolink in the templates.

Change-Id: Ic1272440476146b88e7936726698414c4ca52c5f
Resolves: #96608
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73105
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>