[TASK] Add acceptance test reports to Gitlab
Gitlab is able to give a nice overview of tests
and how often a test failed recently which
makes it easier to investigate flaky/failed
tests. Also it takes away the pain of searching
for the actual failing test in a job because it only
displays information of the failed test
(no bootstrap log details and such)

Codeception AC test reports have been added to the
Gitlab GUI using the --xml flag to generate a Gitlab
readable report format and it has been added to the
jobs artifacts:reports:junit.

Resolves: #96628
Releases: main, 11.5
Change-Id: Id5ab5adfdc318c0c97143b0f05311f47fd1d37f0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73136
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>