[TASK] Revert "[TASK] Ensure requirejs-loader.js and JavaScriptItemHandler.js are cache invalidatable"
This reverts commit 425f94f553b14fa63668beaa82489316f0f3ea4a.

Reason for revert:
It would have been better to use getStreamlinedFileName()
directly instead of changing the processJsFile() method.
Also JavaScriptRenderer is reinitialized in reset(),
therefore the initialization needs to be done again there.
Last but not least this fix should also be backported to v11
as JavaScriptHandler changed between v11.5.3 and v11.5.4 and
therefore a cache-bust hash is required as well.
In order to fix this in a clean way, this patch is reverted
and will be replaced by a new one.

Releases: main
Resolves: #96664
Reverts: #96606
Change-Id: Iab2095823f32eecb972589b3c13a3e38c1f1709b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73161
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>