[Feature] Add option to recursive fetch data for config.cache This adds a third parameter to the `config.cache` configuration. The parameter is optional set to false by default. Records from any subfolder are considered in the calculation of the cache-lifetime by starttime/endtime.
The idea aims at pages with many nested sysfolders for records.
Specifying all the directories manually can be a lot of work.

Use it by adding a third parameter like in this example:
'''config.cache.123 = tx_my_table:456:1'''
to clear the cache on page 123 when a record found in page 456
or any nested subfolder reaches its starttime/endtime