[TASK] Avoid phpstan 'PHPDoc tag @param has invalid value' errors
Annotations are sometimes:
/** @param $foo */
/** @param $foo MyType */

But should be (at least):
/** @param MyType $foo */

Using argument typehints are the preferred solution.
'@param' docblock annotations are removed if argument
type hints can be set or has been already set.

Two test classes with broken syntax are added to
phpstan exclude configuration as this is for testing
purposes and can not be disabled otherwise.
Having them as ignore patterns should be avoided.

Generated baseline with:
> Build/Scripts/runTests.sh -s phpstanGenerateBaseline

Resolves: #96698
Releases: main, 11.5
Change-Id: I3a8eec5ea8dabf93a19b68a92f5d4bbe7bc2249d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73231
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>