[TASK] Remove broken test from ServiceProviderCompilationPassTest
PHPStan static analysis revealed that the method
exceptionForInvalidFactories() did not contain a proper test
setup, which went undetected as the expected TypeError
was thrown for the invalid test setup.

This could have been fixed but the test actually only
tried to assert whether a callable type declaration is set,
in order to test whether a PHP error is thrown when an
invalid factory is provided. This is useless as such
factory will result in an error once called anyway.

Releases: main, 11.5
Resolves: #96706
Change-Id: Ife52768c82a9c2c394ac1d580e57c990e53a7830
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73249
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>