[TASK] Simplify NewContentElementController view rendering
The NewContentElementController (+content button in page
or list) view handling is odd: The single items are
rendered with a view. Those are channeled through
ModuleTemplate->getDynamicTabMenu() which creates
another view to render the 'tab' panel. That panel
is then put into a third view to render the rest
of the modal content.

The patch streamlines this by avoiding getDynamicTabMenu()
and putting all needed variables into one view instance.

Resolves: #96819
Releases: main
Change-Id: Ib5c6790611286c054722e3652e6d5224c3ec827a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73388
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>