[TASK] Remove confusing doc comments in extbase
Some extbase classes related to FAL have been added
back in 2012 with the doc comment, that classes are
experimental and subject to change. Since the classes
are actively used in the TYPO3 core for nearly 10
years now and are marked as @internal, the comments
about the experimental state can savely be removed.

Resolves: #96855
Releases: main
Change-Id: Ifb98fac3289feeb0a4fa0f2468b50512158340d3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73451
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>