[!!!][FEATURE] Auto-registration of NewContentElementWizard via TCA
This change automatically registers all TCA types
to the new content element wizard, so no pageTS
is needed anymore.

The old pageTS config is still evaluated and
overrides existing TCA config. This way it is
still possible to add additional wizard items
with special default values.

Due to the switch, some groups are renamed and
the structuring has slightly changed as it is
now aligned with the CType switch select box.
Most importantly the group `common` is now
named `default`. A permanent migration is in
place.

Items can still be removed through TCEFORM. This
will remove both, the entry in the CType select
and the wizard item. This is the old behavior.

Since all groups and elements, defined in TCA
are added automatically, the previously `show`
list is no longer evaluated. To remove groups
and elements only in the wizard, a new option
"removeItems" is available.

In addition, labels of TCA for tt_content
have been consolidated into EXT:frontend/Resources/Private/locallang_ttc.xlf

Resolves: #102834
Releases: main
Change-Id: I0f9beacd5abd7a129a139172199ac26cab5ce215
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73488
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>