[BUGFIX] Keep ADMCMD_simUser and ADMCMD_simTime in links
When previewing the frontend for a specific
usergroup (ADMCMD_simUser) or previewing a page
in the future (ADMCMD_simTime) the GET parameters
are now continued to be used when browsing the frontend.

This solves issues when a user-restricted page
or a page visible in the future has a form on it,
and the GET parameters are kept.

Resolves: #94974
Resolves: #96491
Releases: main, 11.5
Change-Id: I0a79749648c0235f0a959a416408540fab2fb4b1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73534
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>