[BUGFIX] Fix system maintainers modal in backend context
With the recent ES6 changes, the EXT:install JavaScript modules
could no longer be loaded into the main backend frame, as the
importmap configuration was not tagged to be included there.

Add the backend.module tag to the EXT:install importmap
in order for  TYPO3/CMS/Install/chosen.jquery.min.js to be
loadable for the system maintainer modal in backend context.

Releases: main
Resolves: #96924
Related: #96323
Change-Id: I6109ce00a331b93c1bef29f0e455fa6be126e387
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73560
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>