[TASK] Avoid additional query in TSFE for BE User Sections
This is a nice one: One of the most underrated features
(pages within Backend User Sections), is checked
within TSFE, if a logged-in backend user has access
to the current page.

This change uses the current "TSFE->page" record
to do the check and avoids 1 additional SQL query.

Resolves: #96934
Releases: main
Change-Id: I3e1c1000ab290e5c6cbcb38b512b7caf4eac5bc2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73569
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>