[BUGFIX] Fix felogin GET/POST redirect mode
When ext:felogin is configured to redirectMode "getpost,login",
the extension does always redirect to a page configured in
TS/Flexform although a valid redirect URL is provided using
`return_url` or `redirect_url` GET parameter. Root cause of the
problem is the wrong assignment of the fluid variable
`redirectURL`, which should be used for GET/POST redirect urls
only.

This patch ensures, that the fluid variable `redirectURL`
is assigned either with GET data from `return_url` or
`redirect_url` if available.

Resolves: #92068
Releases: main, 11.5, 10.4
Change-Id: Iae3ba5c4345b26a1cf9d52801c6d101c5986dbb4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73585
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan P <spo@snk.de>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Stefan P <spo@snk.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>